requirements.txt
numpy
typing_extensions
accelerate
audiomentations
bitsandbytes
cyrtranslit
datasets
dtw-python
envsubst
googletrans>=3.1.0a0
huggingface_hub
jiwer>=3.0.5
langid
lang-trans
Levenshtein
librosa
loralib
matplotlib<=3.7.4
num2words
onnxruntime
openai-whisper
peft
psutil
praat-textgrids
PyAudio
py3nvml
pyannote.audio
pyannote.pipeline
pyarabic
pyctcdecode
pypi-kenlm
python-magic
python-slugify
so-vits-svc-fork
soxbindings
speechbrain
tensorboard
text2num
torch
torchaudio
torchvision
transformers
xmltodict
vosk
websockets